requirements.txt
fastmcp>=0.1.0
httpx>=0.23.0
mcp>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=0.21.0
mistralai>=0.0.11
python-socketio>=5.8.0
eventlet>=0.33.3
python-engineio>=4.5.1
fastapi>=0.95.0
uvicorn>=0.21.1
